Make new containers, separate the components into files
