+++ public/index.html
@@ -1,4 +1,4 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="utf-8">
@@ -10,7 +10,9 @@
     <meta property="og:title" content="ICO Transparency Monitor">
     <meta property="og:url" content="https://icomonitor.io/">
     <meta property="og:image" content="https://icomonitor.io/social_logo.jpg">
-    <meta property="og:description" content="A go to place for investors and journalists to learn whether a given ICO is transparent and can be trusted. The ICO Monitor serves as a transparency benchmark for the community to better evaluate offerings and make informed decisions based on objective metrics."/>
+    <meta property="og:description" content="A go to place for investors and journalists to learn whether
+     a given ICO is transparent and can be trusted. The ICO Monitor serves as a transparency benchmark for
+      the community to better evaluate offerings and make informed decisions based on objective metrics."/>
     <title>ICO Transparency Monitor</title>
   </head>
   <body>
@@ -18,7 +20,7 @@
       You need to enable JavaScript to run this app.
     </noscript>
     <div id="root"></div>
-    <script src="main.js"></script>
+    <script type="text/jsx" src="main.js"></script>
     <script>
       window.onhashchange = function(e) {
         if(location.hash === "#/")
