+++ src/components/TokenIssued.js
@@ -1,0 +1,44 @@
+import React from 'react';
+import { formatNumber } from '../utils';
+
+export default ({ tokenIssued, tokensOverflow, totalInvestors, totalTransactions }) => (
+  <div >
+    <h3 className="title">
+      <span
+        className="tooltip"
+        data-tip="This section shows ICO activity over time.<br/> Were most transactions done
+         first day? Were there any spikes? Activity at the end?"
+      >
+        Tokens issuance
+      </span>
+    </h3>
+    <div className="stats">
+      <table>
+        <tbody>
+          <tr>
+            <th>Number of Investors</th>
+            <td>{formatNumber(totalInvestors, 0)}</td>
+          </tr>
+          <tr>
+            <th>Number of Transactions</th>
+            <td>{formatNumber(totalTransactions, 0)}</td>
+          </tr>
+          <tr>
+            <th>Number of tokens created during the ICO</th>
+            <td>{formatNumber(tokenIssued)}</td>
+          </tr>
+          {
+            parseFloat(tokensOverflow.toFixed(2)) !== 0 &&
+            <tr>
+              <th>
+                Number of tokens created outside of ICO<br />
+                <i>*those tokens are not part of results below*</i>
+              </th>
+              <td>{formatNumber(tokensOverflow)}</td>
+            </tr>
+          }
+        </tbody>
+      </table>
+    </div>
+  </div>
+);
